XAI_Anomly_Detection_Component/requirements.txt
pytest-cov
numpy
pandas
scikit-learn
shap
lime
matplotlib
plotly
xgboost
joblib